ML-Ops/app/requirements.txt
flask==3.1.0
numpy==1.26.3
pillow==10.2.0
torch==2.5.1
torchvision==0.20.1